Aligned advanced and session google pay component by introducing a common base component
