+++ lib/src/components/google_pay/google_pay_session_component.dart
@@ -1,188 +1,31 @@
-import 'dart:async';
-
 import 'package:adyen_checkout/src/common/model/payment_result.dart';
-import 'package:adyen_checkout/src/components/component_flutter_api.dart';
-import 'package:adyen_checkout/src/components/component_platform_api.dart';
-import 'package:adyen_checkout/src/components/google_pay/model/google_pay_component_configuration.dart';
+import 'package:adyen_checkout/src/components/google_pay/base_google_pay_component.dart';
 import 'package:adyen_checkout/src/generated/platform_api.g.dart';
 import 'package:adyen_checkout/src/logging/adyen_logger.dart';
-import 'package:adyen_checkout/src/util/dto_mapper.dart';
-import 'package:adyen_checkout/src/util/sdk_version_number_provider.dart';
-import 'package:flutter/material.dart';
-import 'package:pay/pay.dart';
 
-class GooglePaySessionComponent extends StatefulWidget {
-  final String googlePayPaymentMethod;
-  final GooglePayComponentConfiguration googlePayComponentConfiguration;
-  final Function(PaymentResult) onPaymentResult;
-  final GooglePayButtonTheme theme;
-  final GooglePayButtonType type;
-  final int cornerRadius;
-  final double width;
-  final double height;
-  final Function()? onUnavailable;
-  final Widget? unavailableWidget;
-  final Widget? loadingIndicator;
-  final AdyenLogger adyenLogger;
+class GooglePaySessionComponent extends BaseGooglePayComponent {
+  @override
   final String componentId = "GOOGLE_PAY_SESSION_COMPONENT";
 
   GooglePaySessionComponent({
     super.key,
-    required this.googlePayPaymentMethod,
-    required this.googlePayComponentConfiguration,
-    required this.onPaymentResult,
-    required this.theme,
-    required this.type,
-    required this.cornerRadius,
-    required this.width,
-    required this.height,
-    this.onUnavailable,
-    this.unavailableWidget,
-    this.loadingIndicator,
+    required super.googlePayPaymentMethod,
+    required super.googlePayComponentConfiguration,
+    required super.onPaymentResult,
+    required super.theme,
+    required super.type,
+    required super.cornerRadius,
+    required super.width,
+    required super.height,
+    super.loadingIndicator,
+    super.onUnavailable,
+    super.unavailableWidget,
     AdyenLogger? adyenLogger,
-  }) : adyenLogger = adyenLogger ?? AdyenLogger.instance;
+  });
 
   @override
-  State<GooglePaySessionComponent> createState() =>
-      _GooglePaySessionComponentState();
-}
-
-class _GooglePaySessionComponentState extends State<GooglePaySessionComponent> {
-  final SdkVersionNumberProvider _sdkVersionNumberProvider =
-      SdkVersionNumberProvider.instance;
-  final ComponentFlutterApi _componentFlutterApi = ComponentFlutterApi.instance;
-  final ComponentPlatformApi _componentPlatformApi =
-      ComponentPlatformApi.instance;
-  final ValueNotifier<bool> _isButtonClickable = ValueNotifier<bool>(true);
-  final ValueNotifier<bool> _isLoading = ValueNotifier<bool>(false);
-
-  @override
-  void initState() {
-    _componentFlutterApi.componentCommunicationStream.stream
-        .where((communicationModel) =>
-            communicationModel.componentId == widget.componentId)
-        .listen(_handleComponentCommunication);
-
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder(
-      future: _isGooglePaySupported(),
-      builder: (
-        BuildContext context,
-        AsyncSnapshot<InstantPaymentSetupResultDTO> snapshot,
-      ) {
-        if (snapshot.connectionState == ConnectionState.done) {
-          if (_isGooglePaySupportedOnDevice(snapshot)) {
-            return _buildGooglePayOrLoadingContainer(snapshot);
-          } else {
-            widget.onUnavailable?.call();
-            return widget.unavailableWidget ?? const SizedBox.shrink();
-          }
-        }
-
-        return widget.loadingIndicator ?? const SizedBox.shrink();
-      },
-    );
-  }
-
-  @override
-  void dispose() {
-    _componentPlatformApi.onDispose(widget.componentId);
-    _componentFlutterApi.dispose();
-    _isButtonClickable.dispose();
-    super.dispose();
-  }
-
-  bool _isGooglePaySupportedOnDevice(
-      AsyncSnapshot<InstantPaymentSetupResultDTO> snapshot) {
-    return snapshot.data?.instantPaymentType == InstantPaymentType.googlePay &&
-        snapshot.data?.isSupported == true;
-  }
-
-  Widget _buildGooglePayOrLoadingContainer(
-      AsyncSnapshot<InstantPaymentSetupResultDTO> snapshot) {
-    return ValueListenableBuilder(
-      valueListenable: _isLoading,
-      builder: (BuildContext context, value, Widget? child) {
-        if (value == true) {
-          return widget.loadingIndicator ?? const SizedBox.shrink();
-        } else {
-          return _buildGooglePayButton(snapshot);
-        }
-      },
-    );
-  }
-
-  Widget _buildGooglePayButton(
-      AsyncSnapshot<InstantPaymentSetupResultDTO> snapshot) {
-    final String allowedPaymentMethods =
-        snapshot.data?.resultData.toString() ?? "[]";
-    final Widget googlePayButton =
-        _buildRawGooglePayButton(PaymentConfiguration.fromJsonString(
-      '''{
-        "provider": "google_pay",
-        "data": {
-          "apiVersion": 2,
-          "apiVersionMinor": 0,
-          "allowedPaymentMethods": $allowedPaymentMethods
-        }}''',
-    ));
-
-    return SizedBox(
-      width: widget.width,
-      height: widget.height,
-      child: ValueListenableBuilder(
-        valueListenable: _isButtonClickable,
-        builder: (BuildContext context, value, Widget? child) {
-          return IgnorePointer(
-            ignoring: value == false,
-            child: googlePayButton,
-          );
-        },
-      ),
-    );
-  }
-
-  RawGooglePayButton _buildRawGooglePayButton(
-      PaymentConfiguration paymentConfiguration) {
-    return RawGooglePayButton(
-      paymentConfiguration: paymentConfiguration,
-      onPressed: onPressed,
-      cornerRadius: widget.cornerRadius,
-      theme: widget.theme,
-      type: widget.type,
-    );
-  }
-
-  void onPressed() {
-    _isButtonClickable.value = false;
-    _componentPlatformApi.onInstantPaymentPressed(
-      InstantPaymentType.googlePay,
-      widget.componentId,
-    );
-  }
-
-  Future<InstantPaymentSetupResultDTO> _isGooglePaySupported() async {
-    final String versionNumber =
-        await _sdkVersionNumberProvider.getSdkVersionNumber();
-    final InstantPaymentConfigurationDTO
-        instantPaymentComponentConfigurationDTO =
-        widget.googlePayComponentConfiguration.toDTO(
-      versionNumber,
-      InstantPaymentType.googlePay,
-    );
-    return await _componentPlatformApi.isInstantPaymentSupportedByPlatform(
-      instantPaymentComponentConfigurationDTO,
-      widget.googlePayPaymentMethod,
-      widget.componentId,
-    );
-  }
-
-  void _handleComponentCommunication(event) {
-    _isButtonClickable.value = true;
+  void handleComponentCommunication(event) {
+    isButtonClickable.value = true;
     if (event.type case ComponentCommunicationType.result) {
       _onResult(event);
     } else if (event.type case ComponentCommunicationType.error) {
@@ -192,19 +35,18 @@ class _GooglePaySessionComponentState extends State<GooglePaySessionComponent> {
     }
   }
 
-  void _onLoading() => _isLoading.value = true;
+  void _onLoading() => isLoading.value = true;
 
   void _onResult(ComponentCommunicationModel event) {
-    _isLoading.value = false;
+    isLoading.value = false;
     String resultCode = event.paymentResult?.resultCode ?? "";
-    widget.adyenLogger
-        .print("Google pay session flow result code: $resultCode");
-    widget.onPaymentResult(PaymentAdvancedFinished(resultCode: resultCode));
+    adyenLogger.print("Google pay session flow result code: $resultCode");
+    onPaymentResult(PaymentAdvancedFinished(resultCode: resultCode));
   }
 
   void _onError(ComponentCommunicationModel event) {
-    _isLoading.value = false;
+    isLoading.value = false;
     String errorMessage = event.data as String;
-    widget.onPaymentResult(PaymentError(reason: errorMessage));
+    onPaymentResult(PaymentError(reason: errorMessage));
   }
 }
